laucnh a shiny app

laucnh a shiny app

Usage

launch_app(data_folder_address, ...)

Arguments

data_folder_address

address of the folder that your data folders are. for example if you have two sets of data such as data1 and data2 and they are in mydata folder then your data_folder_address should be "./mydata"

...

other variables

Value

open a shiny app